requirements.txt
Flask==1.1.1
Jinja2==2.11.1
matplotlib
numpy
pandas
python-dotenv
mplfinance
yfinance
ffn
pandas_datareader
cvxopt
seaborn

